backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
easyocr==1.7.2
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.2.0
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
PyMuPDF==1.25.2
pytesseract==0.3.13
python-bidi==0.6.3
python-dotenv==1.0.1
PyYAML==6.0.2
scikit-image==0.25.1
scipy==1.15.1
setuptools==75.8.0
shapely==2.0.7
sympy==1.13.1
tifffile==2025.1.10
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
Werkzeug==3.1.3
